# api/routes/swagger.json
{
    "swagger" : "2.0",
    "info" : {
      "version": "2.3.1",
          "title": "REST-API Documentation \"โปรแกรมจัดตารางสอน\"",
          "description": "เอกสารรวบรวมรูปแบบการส่งข้อมูลโปรแกรมจัดตารางสอน",
          "license": {
              "name": "SurapatNJ & Suhaimee24",
              "url": "https://github.com/suhaimee24/web-timetable"
          }
    },
    "paths" : {
    "/API/login" : {
      "post" : {
        "tags":["Authentication"],
        "consumes" : [ "application/json" ],
        "produces" : [ "text/html" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Model15"
          },
          "x-examples" : {
            "application/json" : "{\n    \"username\" : \"admin\",\n    \"password\" : \"1234\"\n}"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Definition generated from Swagger Inspector",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
        "/API/{table_name}" : {
                "get" : {
                    "tags":["CALL"],
                "produces" : [ "application/json" ],
                "parameters" : [ { 
                    "name" : "table_name",
                    "in" : "path",
                    "description" : "List : \nsubject_section\ncurriculum2_subject\ncurriculum2_section\nteach_table\ncurriculum2\nsubject\nteach_table",
                    "required" : true,
                    "type" : "string"
                } ],
                "responses" : {
                    "200" : {
                    "description" : "OK",
                    "schema" : {
                        "$ref" : "#/definitions/Model8"
                    }
                    }
                }
                }
            },
        "/API/subject_section/" : {
            "post" : {
                "tags":["subject_section"],
            "consumes" : [ "application/json" ],
            "produces" : [ "text/html" ],
            "parameters" : [ {
                "in" : "body",
                "name" : "body",
                "required" : false,
                "schema" : {
                "$ref" : "#/definitions/Model4"
                },
                "x-examples" : {
                "application/json" : "{\n    \"subject_id\" : \"01000002\" ,\n    \"subject_section\" : 1 ,\n    \"teach_hr\" : 3 ,\n    \"subject_section_student_amount\" : 40 ,\n    \"teach_day\" : \"tue\",\n    \"teach_time\" : \"13:00:00\" ,\n    \"teach_time2\" : \"16:15:00\",\n    \"lect_or_prac\" : \"l\",\n    \"break_time\" : 15\n}"
                }
            } ],
            "responses" : {
                "200" : {
                "description" : "OK",
                "schema" : {
                    "type" : "string"
                }
                }
            }
            },
            "put" : {
                "tags":["subject_section"],
            "consumes" : [ "application/json" ],
            "produces" : [ "text/html" ],
            "parameters" : [ {
                "in" : "body",
                "name" : "body",
                "required" : false,
                "schema" : {
                "$ref" : "#/definitions/Model5"
                },
                "x-examples" : {
                "application/json" : "{\n    \"subject_id\" : \"01000002\" ,\n    \"subject_section\" : 1 ,\n    \"teach_hr\" : 3 ,\n    \"subject_section_student_amount\" : 40 ,\n    \"teach_day\" : \"mon\",\n    \"teach_time\" : \"13:00:00\" ,\n    \"teach_time2\" : \"16:15:00\",\n    \"lect_or_prac\" : \"l\",\n    \"break_time\" : 15\n}"
                }
            } ],
            "responses" : {
                "200" : {
                "description" : "OK",
                "schema" : {
                    "type" : "string"
                }
                }
            }
            },
            "delete" : {
                "tags":["subject_section"],
            "consumes" : [ "application/json" ],
            "produces" : [ "text/html" ],
            "parameters" : [ {
                "in" : "body",
                "name" : "body",
                "required" : false,
                "schema" : {
                "$ref" : "#/definitions/Model6"
                },
                "x-examples" : {
                "application/json" : "{\r\n    \"subject_id\" : \"01000002\" ,\r\n    \"subject_section\" : 1\r\n}"
                }
            } ],
            "responses" : {
                "200" : {
                "description" : "OK",
                "schema" : {
                    "type" : "string"
                }
                }
            }
            }
        },
        "/API/curriculum2_subject/" : {
            "post" : {
                "tags":["curriculum2_subject"],
            "consumes" : [ "application/json" ],
            "produces" : [ "text/html" ],
            "parameters" : [ {
                "in" : "body",
                "name" : "body",
                "required" : false,
                "schema" : {
                "$ref" : "#/definitions/Model1"
                },
                "x-examples" : {
                "application/json" : "{\n        \"curr2_id\": \"06\",\n        \"subject_id\": \"01000002\",\n        \"semester\": 1\n   }"
                }
            } ],
            "responses" : {
                "200" : {
                "description" : "OK",
                "schema" : {
                    "type" : "string"
                }
                }
            }
            },
            "delete" : {
                "tags":["curriculum2_subject"],
            "consumes" : [ "application/json" ],
            "produces" : [ "text/html" ],
            "parameters" : [ {
                "in" : "body",
                "name" : "body",
                "required" : false,
                "schema" : {
                "$ref" : "#/definitions/Model7"
                },
                "x-examples" : {
                "application/json" : "{\n        \"curr2_id\": \"06\",\n        \"subject_id\": \"01000002\",\n        \"semester\": 1\n   }"
                }
            } ],
            "responses" : {
                "200" : {
                "description" : "OK",
                "schema" : {
                    "type" : "string"
                }
                }
            }
            }
        },
        "/API/curriculum2_section/" : {
            "post" : {
                "tags":["curriculum2_section"],
            "consumes" : [ "application/json" ],
            "produces" : [ "text/html" ],
            "parameters" : [ {
                "in" : "body",
                "name" : "body",
                "required" : false,
                "schema" : {
                "$ref" : "#/definitions/Model0"
                },
                "x-examples" : {
                "application/json" : "{    \r\n    \"curr2_id\" : \"07\" ,\r\n    \"curr2_section\":1,\r\n    \"curr2_section_student_amount\":1\r\n}"
                }
            } ],
            "responses" : {
                "200" : {
                "description" : "OK",
                "schema" : {
                    "type" : "string"
                }
                }
            }
            },
            "put" : {
                "tags":["curriculum2_section"],
            "consumes" : [ "application/json" ],
            "produces" : [ "text/html" ],
            "parameters" : [ {
                "in" : "body",
                "name" : "body",
                "required" : false,
                "schema" : {
                "$ref" : "#/definitions/Model2"
                },
                "x-examples" : {
                "application/json" : "{    \r\n    \"curr2_id\" : \"07\" ,\r\n    \"curr2_section\":1,\r\n    \"curr2_section_student_amount\":40\r\n}"
                }
            } ],
            "responses" : {
                "200" : {
                "description" : "OK",
                "schema" : {
                    "type" : "string"
                }
                }
            }
            },
            "delete" : {
                "tags":["curriculum2_section"],
            "consumes" : [ "application/json" ],
            "produces" : [ "text/html" ],
            "parameters" : [ {
                "in" : "body",
                "name" : "body",
                "required" : false,
                "schema" : {
                "$ref" : "#/definitions/Model3"
                },
                "x-examples" : {
                "application/json" : "{    \r\n    \"curr2_id\" : \"07\" ,\r\n    \"curr2_section\":1\r\n}"
                }
            } ],
            "responses" : {
                "200" : {
                "description" : "OK",
                "schema" : {
                    "type" : "string"
                }
                }
            }
            }
        },
        "/API/curriculum2" : {
          "post" : {
            "tags":["curriculum2"],
            "consumes" : [ "application/json" ],
            "produces" : [ "text/html" ],
            "parameters" : [ {
              "in" : "body",
              "name" : "body",
              "required" : false,
              "schema" : {
                "$ref" : "#/definitions/Model12"
              },
              "x-examples" : {
                "application/json" : "{\n        \"curr2_id\": \"63\",\n        \"dept_id\": \"13\",\n        \"curr2_tname\": \"TEST)\",\n        \"curr2_ename\": \"-\"\n    }"
              }
            } ],
            "responses" : {
              "200" : {
                "description" : "OK",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "put" : {
            "tags":["curriculum2"],
            "consumes" : [ "application/json" ],
            "produces" : [ "text/html" ],
            "parameters" : [ {
              "in" : "body",
              "name" : "body",
              "required" : false,
              "schema" : {
                "$ref" : "#/definitions/Model13"
              },
              "x-examples" : {
                "application/json" : "{\n        \"curr2_id\": \"63\",\n        \"dept_id\": \"13\",\n        \"curr2_tname\": \"TEST2\",\n        \"curr2_ename\": \"-\"\n    }"
              }
            } ],
            "responses" : {
              "200" : {
                "description" : "OK",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "delete" : {
            "tags":["curriculum2"],
            "consumes" : [ "application/json" ],
            "produces" : [ "text/html" ],
            "parameters" : [ {
              "in" : "body",
              "name" : "body",
              "required" : false,
              "schema" : {
                "$ref" : "#/definitions/Model14"
              },
              "x-examples" : {
                "application/json" : "{\n        \"curr2_id\": \"63\"\n    }"
              }
            } ],
            "responses" : {
              "200" : {
                "description" : "OK",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "/API/teach_table/" : {
          "post" : {
            "tags":["teach_table"],
            "consumes" : [ "application/json" ],
            "parameters" : [ {
              "in" : "body",
              "name" : "body",
              "required" : false,
              "schema" : {
                "$ref" : "#/definitions/Model16"
              },
              "x-examples" : {
                "application/json" : "{\n    \"subject_id\":\"01006012\",  \n    \"subject_section\":\"5\",\n    \"curr2_section\":\"1\",\n    \"teach_time\":\"08:00:00\",\n    \"semester\":\"1\",\n    \"year\":\"2020\",\n    \"curr2_id\":\"02\",\n    \"teach_day\":\"mon\",\n    \"teach_time2\":\"10:00:00\",\n    \"lect_or_prac\":\"l\",\n    \"break_time\":\"null\"\n   }"
              }
            } ],
            "responses" : {
              "default" : {
                "description" : "Definition generated from Swagger Inspector"
              }
            },
            "security" : [ {
              "Secured" : [ ]
            } ]
          },
          "put" : {
            "tags":["teach_table"],
            "consumes" : [ "application/json" ],
            "produces" : [ "text/html" ],
            "parameters" : [ {
              "in" : "body",
              "name" : "body",
              "required" : false,
              "schema" : {
                "$ref" : "#/definitions/Model17"
              },
              "x-examples" : {
                "application/json" : "{\n    \"subject_id\":\"01006012\",  \n    \"subject_section\":\"5\",\n    \"curr2_section\":\"1\",\n    \"teach_time\":\"'08:00:00'\",\n    \"semester\":\"1\",\n    \"year\":\"2020\",\n    \"curr2_id\":\"02\",\n    \"teach_day\":\"'mon'\",\n    \"teach_time2\":\"'10:00:00'\",\n    \"lect_or_prac\":\"'p'\",\n    \"break_time\":\"null\"\n   }"
              }
            } ],
            "responses" : {
              "200" : {
                "description" : "Definition generated from Swagger Inspector",
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "security" : [ {
              "Secured" : [ ]
            } ]
          },
          "delete" : {
            "tags":["teach_table"],
            "consumes" : [ "application/json" ],
            "produces" : [ "text/html" ],
            "parameters" : [ {
              "in" : "body",
              "name" : "body",
              "required" : false,
              "schema" : {
                "$ref" : "#/definitions/Model18"
              },
              "x-examples" : {
                "application/json" : "{\n    \"subject_id\":\"01006012\",  \n    \"subject_section\":\"5\",\n    \"curr2_section\":\"1\",\n    \"teach_time\":\"'08:00:00'\",\n    \"semester\":\"1\",\n    \"year\":\"2020\"\n   }"
              }
            } ],
            "responses" : {
              "200" : {
                "description" : "Definition generated from Swagger Inspector",
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "security" : [ {
              "Secured" : [ ]
            } ]
          }
        }
    },
    "definitions" : {
      "Model0" : {
        "properties" : {
          "curr2_id" : {
            "type" : "string"
          },
          "curr2_section" : {
            "type" : "integer",
            "format" : "int32"
          },
          "curr2_section_student_amount" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "Model1" : {
        "properties" : {
          "curr2_id" : {
            "type" : "string"
          },
          "subject_id" : {
            "type" : "string"
          },
          "semester" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "Model2" : {
        "properties" : {
          "curr2_id" : {
            "type" : "string"
          },
          "curr2_section" : {
            "type" : "integer",
            "format" : "int32"
          },
          "curr2_section_student_amount" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "Model3" : {
        "properties" : {
          "curr2_id" : {
            "type" : "string"
          },
          "curr2_section" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "Model4" : {
        "properties" : {
          "subject_id" : {
            "type" : "string"
          },
          "subject_section" : {
            "type" : "integer",
            "format" : "int32"
          },
          "teach_hr" : {
            "type" : "integer",
            "format" : "int32"
          },
          "subject_section_student_amount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "teach_day" : {
            "type" : "string"
          },
          "teach_time" : {
            "type" : "string"
          },
          "teach_time2" : {
            "type" : "string"
          },
          "lect_or_prac" : {
            "type" : "string"
          },
          "break_time" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "Model5" : {
        "properties" : {
          "subject_id" : {
            "type" : "string"
          },
          "subject_section" : {
            "type" : "integer",
            "format" : "int32"
          },
          "teach_hr" : {
            "type" : "integer",
            "format" : "int32"
          },
          "subject_section_student_amount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "teach_day" : {
            "type" : "string"
          },
          "teach_time" : {
            "type" : "string"
          },
          "teach_time2" : {
            "type" : "string"
          },
          "lect_or_prac" : {
            "type" : "string"
          },
          "break_time" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "Model6" : {
        "properties" : {
          "subject_id" : {
            "type" : "string"
          },
          "subject_section" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "Model7" : {
        "properties" : {
          "curr2_id" : {
            "type" : "string"
          },
          "subject_id" : {
            "type" : "string"
          },
          "semester" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "Model8" : {
        "type" : "array"
      },
      "Array" : {
        "properties" : {
          "curr2_id" : {
            "type" : "string"
          },
          "subject_id" : {
            "type" : "string"
          },
          "semester" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "Model11_Array" : {
        "properties" : {
          "curr2_id" : {
            "type" : "string"
          },
          "curr2_section" : {
            "type" : "integer",
            "format" : "int32"
          },
          "curr2_section_student_amount" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "Model12" : {
        "properties" : {
          "curr2_id" : {
            "type" : "string"
          },
          "dept_id" : {
            "type" : "string"
          },
          "curr2_tname" : {
            "type" : "string"
          },
          "curr2_ename" : {
            "type" : "string"
          }
        }
      },
      "Model13" : {
        "properties" : {
          "curr2_id" : {
            "type" : "string"
          },
          "dept_id" : {
            "type" : "string"
          },
          "curr2_tname" : {
            "type" : "string"
          },
          "curr2_ename" : {
            "type" : "string"
          }
        }
      },
      "Model14" : {
        "properties" : {
          "curr2_id" : {
            "type" : "string"
          }
        }
      },
      "Model15" : {
        "properties" : {
          "username" : {
            "type" : "string"
          },
          "password" : {
            "type" : "string"
          }
        }
      },    
      "Model16" : {
        "properties" : {
          "subject_id" : {
            "type" : "string"
          },
          "subject_section" : {
            "type" : "string"
          },
          "curr2_section" : {
            "type" : "string"
          },
          "teach_time" : {
            "type" : "string"
          },
          "semester" : {
            "type" : "string"
          },
          "year" : {
            "type" : "string"
          },
          "curr2_id" : {
            "type" : "string"
          },
          "teach_day" : {
            "type" : "string"
          },
          "teach_time2" : {
            "type" : "string"
          },
          "lect_or_prac" : {
            "type" : "string"
          },
          "break_time" : {
            "type" : "string"
          }
        }
      },
      "Model17" : {
        "properties" : {
          "subject_id" : {
            "type" : "string"
          },
          "subject_section" : {
            "type" : "string"
          },
          "curr2_section" : {
            "type" : "string"
          },
          "teach_time" : {
            "type" : "string"
          },
          "semester" : {
            "type" : "string"
          },
          "year" : {
            "type" : "string"
          },
          "curr2_id" : {
            "type" : "string"
          },
          "teach_day" : {
            "type" : "string"
          },
          "teach_time2" : {
            "type" : "string"
          },
          "lect_or_prac" : {
            "type" : "string"
          },
          "break_time" : {
            "type" : "string"
          }
        }
      },
      "Model18" : {
        "properties" : {
          "subject_id" : {
            "type" : "string"
          },
          "subject_section" : {
            "type" : "string"
          },
          "curr2_section" : {
            "type" : "string"
          },
          "teach_time" : {
            "type" : "string"
          },
          "semester" : {
            "type" : "string"
          },
          "year" : {
            "type" : "string"
          }
        }
      }
    }
  }